Revert "[TEST] account 앱 단위 테스트 작성 #91"
