+++ backend/account/serializers.py
@@ -15,7 +15,20 @@ from config.exceptions import (
     InvalidFieldStateException,
     InvalidRefreshTokenException,
 )
-from config.utils import WorkoutLevelChoiceField
+
+
+class WorkoutLevelChoiceField(serializers.ChoiceField):
+    def to_representation(self, obj):
+        for key, val in self._choices.items():
+            if key == int(obj):
+                return val
+
+    def to_internal_value(self, data):
+        # To support inserts with the value
+        for key, val in self._choices.items():
+            if val == data:
+                return key
+        raise InvalidFieldException("난이도가 정확하지 않습니다.")
 
 
 class UserRegistrationSerializer(serializers.ModelSerializer):
